Consolidate RPC utility functions (RIPD-947)
