+++ Builds/VisualStudio2015/RippleD.vcxproj.filters
@@ -3753,51 +3753,27 @@
     <ClCompile Include="..\..\src\ripple\rpc\handlers\WalletSeed.cpp">
       <Filter>ripple\rpc\handlers</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\src\ripple\rpc\impl\AccountFromString.cpp">
-      <Filter>ripple\rpc\impl</Filter>
-    </ClCompile>
-    <ClInclude Include="..\..\src\ripple\rpc\impl\AccountFromString.h">
-      <Filter>ripple\rpc\impl</Filter>
-    </ClInclude>
-    <ClCompile Include="..\..\src\ripple\rpc\impl\GetAccountObjects.cpp">
-      <Filter>ripple\rpc\impl</Filter>
-    </ClCompile>
-    <ClInclude Include="..\..\src\ripple\rpc\impl\GetAccountObjects.h">
-      <Filter>ripple\rpc\impl</Filter>
-    </ClInclude>
     <ClCompile Include="..\..\src\ripple\rpc\impl\Handler.cpp">
       <Filter>ripple\rpc\impl</Filter>
     </ClCompile>
     <ClInclude Include="..\..\src\ripple\rpc\impl\Handler.h">
       <Filter>ripple\rpc\impl</Filter>
     </ClInclude>
-    <ClCompile Include="..\..\src\ripple\rpc\impl\KeypairForSignature.cpp">
-      <Filter>ripple\rpc\impl</Filter>
-    </ClCompile>
     <ClCompile Include="..\..\src\ripple\rpc\impl\LegacyPathFind.cpp">
       <Filter>ripple\rpc\impl</Filter>
     </ClCompile>
     <ClInclude Include="..\..\src\ripple\rpc\impl\LegacyPathFind.h">
       <Filter>ripple\rpc\impl</Filter>
     </ClInclude>
-    <ClCompile Include="..\..\src\ripple\rpc\impl\LookupLedger.cpp">
-      <Filter>ripple\rpc\impl</Filter>
-    </ClCompile>
-    <ClInclude Include="..\..\src\ripple\rpc\impl\LookupLedger.h">
-      <Filter>ripple\rpc\impl</Filter>
-    </ClInclude>
-    <ClCompile Include="..\..\src\ripple\rpc\impl\ParseAccountIds.cpp">
-      <Filter>ripple\rpc\impl</Filter>
-    </ClCompile>
-    <ClInclude Include="..\..\src\ripple\rpc\impl\ParseAccountIds.h">
-      <Filter>ripple\rpc\impl</Filter>
-    </ClInclude>
     <ClCompile Include="..\..\src\ripple\rpc\impl\RPCHandler.cpp">
       <Filter>ripple\rpc\impl</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\src\ripple\rpc\impl\RPCVersion.cpp">
+    <ClCompile Include="..\..\src\ripple\rpc\impl\RPCHelpers.cpp">
       <Filter>ripple\rpc\impl</Filter>
     </ClCompile>
+    <ClInclude Include="..\..\src\ripple\rpc\impl\RPCHelpers.h">
+      <Filter>ripple\rpc\impl</Filter>
+    </ClInclude>
     <ClCompile Include="..\..\src\ripple\rpc\impl\Status.cpp">
       <Filter>ripple\rpc\impl</Filter>
     </ClCompile>
@@ -3810,25 +3786,13 @@
     <ClInclude Include="..\..\src\ripple\rpc\impl\Tuning.h">
       <Filter>ripple\rpc\impl</Filter>
     </ClInclude>
-    <ClCompile Include="..\..\src\ripple\rpc\impl\Utilities.cpp">
-      <Filter>ripple\rpc\impl</Filter>
-    </ClCompile>
-    <ClInclude Include="..\..\src\ripple\rpc\impl\Utilities.h">
-      <Filter>ripple\rpc\impl</Filter>
-    </ClInclude>
     <ClInclude Include="..\..\src\ripple\rpc\InternalHandler.h">
-      <Filter>ripple\rpc</Filter>
-    </ClInclude>
-    <ClInclude Include="..\..\src\ripple\rpc\KeypairForSignature.h">
       <Filter>ripple\rpc</Filter>
     </ClInclude>
     <ClInclude Include="..\..\src\ripple\rpc\RipplePathFind.h">
       <Filter>ripple\rpc</Filter>
     </ClInclude>
     <ClInclude Include="..\..\src\ripple\rpc\RPCHandler.h">
-      <Filter>ripple\rpc</Filter>
-    </ClInclude>
-    <ClInclude Include="..\..\src\ripple\rpc\RPCVersion.h">
       <Filter>ripple\rpc</Filter>
     </ClInclude>
     <ClInclude Include="..\..\src\ripple\rpc\Status.h">
