+++ src/ripple/rpc/handlers/AccountInfo.cpp
@@ -26,10 +26,8 @@
 #include <ripple/protocol/Indexes.h>
 #include <ripple/protocol/JsonFields.h>
 #include <ripple/protocol/types.h>
-#include <ripple/rpc/impl/Utilities.h>
 #include <ripple/rpc/Context.h>
-#include <ripple/rpc/impl/AccountFromString.h>
-#include <ripple/rpc/impl/LookupLedger.h>
+#include <ripple/rpc/impl/RPCHelpers.h>
 
 namespace ripple {
