+++ src/ripple/rpc/handlers/AccountOffers.cpp
@@ -21,15 +21,14 @@
 #include <ripple/app/main/Application.h>
 #include <ripple/json/json_value.h>
 #include <ripple/ledger/ReadView.h>
+#include <ripple/ledger/View.h>
 #include <ripple/net/RPCErr.h>
 #include <ripple/protocol/ErrorCodes.h>
 #include <ripple/protocol/JsonFields.h>
 #include <ripple/resource/Fees.h>
 #include <ripple/rpc/Context.h>
-#include <ripple/rpc/impl/AccountFromString.h>
-#include <ripple/rpc/impl/LookupLedger.h>
+#include <ripple/rpc/impl/RPCHelpers.h>
 #include <ripple/rpc/impl/Tuning.h>
-#include <ripple/rpc/impl/Utilities.h>
 
 namespace ripple {
