+++ src/ripple/rpc/handlers/BookOffers.cpp
@@ -28,8 +28,7 @@
 #include <ripple/protocol/types.h>
 #include <ripple/resource/Fees.h>
 #include <ripple/rpc/Context.h>
-#include <ripple/rpc/impl/LookupLedger.h>
-#include <ripple/rpc/impl/Utilities.h>
+#include <ripple/rpc/impl/RPCHelpers.h>
 
 namespace ripple {
