+++ src/ripple/rpc/handlers/LedgerHandler.cpp
@@ -26,7 +26,7 @@
 #include <ripple/protocol/ErrorCodes.h>
 #include <ripple/protocol/JsonFields.h>
 #include <ripple/resource/Fees.h>
-#include <ripple/rpc/impl/LookupLedger.h>
+#include <ripple/rpc/impl/RPCHelpers.h>
 #include <ripple/server/Role.h>
 
 namespace ripple {
