+++ src/ripple/rpc/handlers/LedgerHeader.cpp
@@ -22,7 +22,7 @@
 #include <ripple/basics/strHex.h>
 #include <ripple/ledger/ReadView.h>
 #include <ripple/protocol/JsonFields.h>
-#include <ripple/rpc/impl/LookupLedger.h>
+#include <ripple/rpc/impl/RPCHelpers.h>
 
 namespace ripple {
