+++ src/ripple/rpc/handlers/NoRippleCheck.cpp
@@ -27,10 +27,8 @@
 #include <ripple/protocol/JsonFields.h>
 #include <ripple/protocol/TxFlags.h>
 #include <ripple/rpc/Context.h>
-#include <ripple/rpc/impl/AccountFromString.h>
-#include <ripple/rpc/impl/LookupLedger.h>
+#include <ripple/rpc/impl/RPCHelpers.h>
 #include <ripple/rpc/impl/Tuning.h>
-#include <ripple/rpc/impl/Utilities.h>
 
 namespace ripple {
