+++ src/ripple/rpc/handlers/Subscribe.cpp
@@ -28,7 +28,7 @@
 #include <ripple/protocol/ErrorCodes.h>
 #include <ripple/protocol/JsonFields.h>
 #include <ripple/resource/Fees.h>
-#include <ripple/rpc/impl/ParseAccountIds.h>
+#include <ripple/rpc/impl/RPCHelpers.h>
 #include <ripple/rpc/Context.h>
 #include <ripple/server/Role.h>
