+++ src/ripple/rpc/handlers/Version.h
@@ -20,7 +20,7 @@
 #ifndef RIPPLED_RIPPLE_RPC_HANDLERS_VERSION_H
 #define RIPPLED_RIPPLE_RPC_HANDLERS_VERSION_H
 
-#include <ripple/rpc/RPCVersion.h>
+#include <ripple/rpc/impl/RPCHelpers.h>
 
 namespace ripple {
 namespace RPC {
