--- src/ripple/rpc/impl/KeypairForSignature.cpp
@@ -1,199 +1,0 @@
-//------------------------------------------------------------------------------
-/*
-    This file is part of rippled: https://github.com/ripple/rippled
-    Copyright (c) 2012-2015 Ripple Labs Inc.
-
-    Permission to use, copy, modify, and/or distribute this software for any
-    purpose  with  or without fee is hereby granted, provided that the above
-    copyright notice and this permission notice appear in all copies.
-
-    THE  SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES
-    WITH  REGARD  TO  THIS  SOFTWARE  INCLUDING  ALL  IMPLIED  WARRANTIES  OF
-    MERCHANTABILITY  AND  FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR
-    ANY  SPECIAL ,  DIRECT, INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES
-    WHATSOEVER  RESULTING  FROM  LOSS  OF USE, DATA OR PROFITS, WHETHER IN AN
-    ACTION  OF  CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF
-    OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
-*/
-//==============================================================================
-
-#include <BeastConfig.h>
-#include <ripple/rpc/KeypairForSignature.h>
-#include <ripple/app/misc/NetworkOPs.h>
-#include <ripple/net/RPCErr.h>
-#include <ripple/protocol/ErrorCodes.h>
-#include <ripple/protocol/JsonFields.h>
-
-namespace ripple {
-namespace RPC {
-
-boost::optional<Seed>
-getSeedFromRPC (Json::Value const& params, Json::Value& error)
-{
-    // The array should be constexpr, but that makes Visual Studio unhappy.
-    static char const* const seedTypes[]
-    {
-        jss::passphrase.c_str(),
-        jss::seed.c_str(),
-        jss::seed_hex.c_str()
-    };
-
-    // Identify which seed type is in use.
-    char const* seedType = nullptr;
-    int count = 0;
-    for (auto t : seedTypes)
-    {
-        if (params.isMember (t))
-        {
-            ++count;
-            seedType = t;
-        }
-    }
-
-    if (count != 1)
-    {
-        error = RPC::make_param_error (
-            "Exactly one of the following must be specified: " +
-            std::string(jss::passphrase) + ", " +
-            std::string(jss::seed) + " or " +
-            std::string(jss::seed_hex));
-        return boost::none;
-    }
-
-    // Make sure a string is present
-    if (! params[seedType].isString())
-    {
-        error = RPC::expected_field_error (seedType, "string");
-        return boost::none;
-    }
-
-    auto const fieldContents = params[seedType].asString();
-
-    // Convert string to seed.
-    boost::optional<Seed> seed;
-
-    if (seedType == jss::seed.c_str())
-        seed = parseBase58<Seed> (fieldContents);
-    else if (seedType == jss::passphrase.c_str())
-        seed = parseGenericSeed (fieldContents);
-    else if (seedType == jss::seed_hex.c_str())
-    {
-        uint128 s;
-
-        if (s.SetHexExact (fieldContents))
-            seed.emplace (Slice(s.data(), s.size()));
-    }
-
-    if (!seed)
-        error = rpcError (rpcBAD_SEED);
-
-    return seed;
-}
-
-std::pair<PublicKey, SecretKey>
-keypairForSignature (Json::Value const& params, Json::Value& error)
-{
-    bool const has_key_type  = params.isMember (jss::key_type);
-
-    // All of the secret types we allow, but only one at a time.
-    // The array should be constexpr, but that makes Visual Studio unhappy.
-    static char const* const secretTypes[]
-    {
-        jss::passphrase.c_str(),
-        jss::secret.c_str(),
-        jss::seed.c_str(),
-        jss::seed_hex.c_str()
-    };
-
-    // Identify which secret type is in use.
-    char const* secretType = nullptr;
-    int count = 0;
-    for (auto t : secretTypes)
-    {
-        if (params.isMember (t))
-        {
-            ++count;
-            secretType = t;
-        }
-    }
-
-    if (count == 0 || secretType == nullptr)
-    {
-        error = RPC::missing_field_error (jss::secret);
-        return { };
-    }
-
-    if (count > 1)
-    {
-        error = RPC::make_param_error (
-            "Exactly one of the following must be specified: " +
-            std::string(jss::passphrase) + ", " +
-            std::string(jss::secret) + ", " +
-            std::string(jss::seed) + " or " +
-            std::string(jss::seed_hex));
-        return { };
-    }
-
-    KeyType keyType = KeyType::secp256k1;
-    boost::optional<Seed> seed;
-
-    if (has_key_type)
-    {
-        if (! params[jss::key_type].isString())
-        {
-            error = RPC::expected_field_error (
-                jss::key_type, "string");
-            return { };
-        }
-
-        keyType = keyTypeFromString (
-            params[jss::key_type].asString());
-
-        if (keyType == KeyType::invalid)
-        {
-            error = RPC::invalid_field_error(jss::key_type);
-            return { };
-        }
-
-        if (secretType == jss::secret.c_str())
-        {
-            error = RPC::make_param_error (
-                "The secret field is not allowed if " +
-                std::string(jss::key_type) + " is used.");
-            return { };
-        }
-
-        seed = getSeedFromRPC (params, error);
-    }
-    else
-    {
-        if (! params[jss::secret].isString())
-        {
-            error = RPC::expected_field_error (
-                jss::secret, "string");
-            return { };
-        }
-
-        seed = parseGenericSeed (
-            params[jss::secret].asString ());
-    }
-
-    if (!seed)
-    {
-        if (!contains_error (error))
-        {
-            error = RPC::make_error (rpcBAD_SEED,
-            RPC::invalid_field_message (secretType));
-        }
-
-        return { };
-    }
-
-    if (keyType != KeyType::secp256k1 && keyType != KeyType::ed25519)
-        LogicError ("keypairForSignature: invalid key type");
-
-    return generateKeyPair (keyType, *seed);
-}
-
-} // RPC
-} // ripple
