--- src/ripple/rpc/impl/LookupLedger.cpp
@@ -1,249 +1,0 @@
-//------------------------------------------------------------------------------
-/*
-    This file is part of rippled: https://github.com/ripple/rippled
-    Copyright (c) 2012-2014 Ripple Labs Inc.
-
-    Permission to use, copy, modify, and/or distribute this software for any
-    purpose  with  or without fee is hereby granted, provided that the above
-    copyright notice and this permission notice appear in all copies.
-
-    THE  SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES
-    WITH  REGARD  TO  THIS  SOFTWARE  INCLUDING  ALL  IMPLIED  WARRANTIES  OF
-    MERCHANTABILITY  AND  FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR
-    ANY  SPECIAL ,  DIRECT, INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES
-    WHATSOEVER  RESULTING  FROM  LOSS  OF USE, DATA OR PROFITS, WHETHER IN AN
-    ACTION  OF  CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF
-    OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
-*/
-//==============================================================================
-
-#include <BeastConfig.h>
-#include <ripple/app/ledger/LedgerMaster.h>
-#include <ripple/app/main/Application.h>
-#include <ripple/app/misc/NetworkOPs.h>
-#include <ripple/basics/Log.h>
-#include <ripple/json/json_value.h>
-#include <ripple/ledger/View.h>
-#include <ripple/rpc/Context.h>
-#include <ripple/rpc/impl/LookupLedger.h>
-#include <ripple/rpc/impl/Tuning.h>
-
-namespace ripple {
-namespace RPC {
-
-namespace {
-
-bool isValidatedOld (LedgerMaster& ledgerMaster, bool standalone)
-{
-    if (standalone)
-        return false;
-
-    return ledgerMaster.getValidatedLedgerAge () >
-        Tuning::maxValidatedLedgerAge;
-}
-
-template <class T>
-Status ledgerFromRequest (T& ledger, Context& context)
-{
-    static auto const minSequenceGap = 10;
-
-    ledger.reset();
-
-    auto& params = context.params;
-    auto& ledgerMaster = context.ledgerMaster;
-
-    auto indexValue = params[jss::ledger_index];
-    auto hashValue = params[jss::ledger_hash];
-
-    // We need to support the legacy "ledger" field.
-    auto& legacyLedger = params[jss::ledger];
-    if (legacyLedger)
-    {
-        if (legacyLedger.asString().size () > 12)
-            hashValue = legacyLedger;
-        else
-            indexValue = legacyLedger;
-    }
-
-    if (hashValue)
-    {
-        if (! hashValue.isString ())
-            return {rpcINVALID_PARAMS, "ledgerHashNotString"};
-
-        uint256 ledgerHash;
-        if (! ledgerHash.SetHex (hashValue.asString ()))
-            return {rpcINVALID_PARAMS, "ledgerHashMalformed"};
-
-        ledger = ledgerMaster.getLedgerByHash (ledgerHash);
-        if (ledger == nullptr)
-            return {rpcLGR_NOT_FOUND, "ledgerNotFound"};
-    }
-    else if (indexValue.isNumeric())
-    {
-        ledger = ledgerMaster.getLedgerBySeq (indexValue.asInt ());
-
-        if (ledger == nullptr)
-        {
-            auto cur = ledgerMaster.getCurrentLedger();
-            if (cur->info().seq == indexValue.asInt())
-                ledger = cur;
-        }
-
-        if (ledger == nullptr)
-            return {rpcLGR_NOT_FOUND, "ledgerNotFound"};
-
-        if (ledger->info().seq > ledgerMaster.getValidLedgerIndex() &&
-            isValidatedOld(ledgerMaster, context.app.config().RUN_STANDALONE))
-        {
-            ledger.reset();
-            return {rpcNO_NETWORK, "InsufficientNetworkMode"};
-        }
-    }
-    else
-    {
-        if (isValidatedOld (ledgerMaster, context.app.config().RUN_STANDALONE))
-            return {rpcNO_NETWORK, "InsufficientNetworkMode"};
-
-        auto const index = indexValue.asString ();
-        if (index == "validated")
-        {
-            ledger = ledgerMaster.getValidatedLedger ();
-            if (ledger == nullptr)
-                return {rpcNO_NETWORK, "InsufficientNetworkMode"};
-
-            assert (! ledger->open());
-        }
-        else
-        {
-            if (index.empty () || index == "current")
-            {
-                ledger = ledgerMaster.getCurrentLedger ();
-                assert (ledger->open());
-            }
-            else if (index == "closed")
-            {
-                ledger = ledgerMaster.getClosedLedger ();
-                assert (! ledger->open());
-            }
-            else
-            {
-                return {rpcINVALID_PARAMS, "ledgerIndexMalformed"};
-            }
-
-            if (ledger == nullptr)
-                return {rpcNO_NETWORK, "InsufficientNetworkMode"};
-
-            if (ledger->info().seq + minSequenceGap <
-                ledgerMaster.getValidLedgerIndex ())
-            {
-                ledger.reset ();
-                return {rpcNO_NETWORK, "InsufficientNetworkMode"};
-            }
-        }
-    }
-
-    return Status::OK;
-}
-
-bool isValidated (LedgerMaster& ledgerMaster, ReadView const& ledger,
-    Application& app)
-{
-    if (ledger.open())
-        return false;
-
-    if (ledger.info().validated)
-        return true;
-
-    auto seq = ledger.info().seq;
-    try
-    {
-        // Use the skip list in the last validated ledger to see if ledger
-        // comes before the last validated ledger (and thus has been
-        // validated).
-        auto hash = ledgerMaster.walkHashBySeq (seq);
-
-        if (!hash || ledger.info().hash != *hash)
-        {
-            // This ledger's hash is not the hash of the validated ledger
-            if (hash)
-            {
-                assert(hash->isNonZero());
-                uint256 valHash = getHashByIndex (seq, app);
-                if (valHash == ledger.info().hash)
-                {
-                    // SQL database doesn't match ledger chain
-                    ledgerMaster.clearLedger (seq);
-                }
-            }
-            return false;
-        }
-    }
-    catch (SHAMapMissingNode const&)
-    {
-        auto stream = app.journal ("RPCHandler").warn();
-        JLOG (stream)
-            << "Missing SHANode " << std::to_string (seq);
-        return false;
-    }
-
-    // Mark ledger as validated to save time if we see it again.
-    ledger.info().validated = true;
-    return true;
-}
-
-} // namespace
-
-// The previous version of the lookupLedger command would accept the
-// "ledger_index" argument as a string and silently treat it as a request to
-// return the current ledger which, while not strictly wrong, could cause a lot
-// of confusion.
-//
-// The code now robustly validates the input and ensures that the only possible
-// values for the "ledger_index" parameter are the index of a ledger passed as
-// an integer or one of the strings "current", "closed" or "validated".
-// Additionally, the code ensures that the value passed in "ledger_hash" is a
-// string and a valid hash. Invalid values will return an appropriate error
-// code.
-//
-// In the absence of the "ledger_hash" or "ledger_index" parameters, the code
-// assumes that "ledger_index" has the value "current".
-//
-// Returns a Json::objectValue.  If there was an error, it will be in that
-// return value.  Otherwise, the object contains the field "validated" and
-// optionally the fields "ledger_hash", "ledger_index" and
-// "ledger_current_index", if they are defined.
-Status lookupLedger (
-    std::shared_ptr<ReadView const>& ledger, Context& context,
-    Json::Value& result)
-{
-    if (auto status = ledgerFromRequest (ledger, context))
-        return status;
-
-    auto& info = ledger->info();
-
-    if (!ledger->open())
-    {
-        result[jss::ledger_hash] = to_string (info.hash);
-        result[jss::ledger_index] = info.seq;
-    }
-    else
-    {
-        result[jss::ledger_current_index] = info.seq;
-    }
-
-    result[jss::validated] = isValidated (context.ledgerMaster, *ledger, context.app);
-    return Status::OK;
-}
-
-Json::Value lookupLedger (
-    std::shared_ptr<ReadView const>& ledger, Context& context)
-{
-    Json::Value result;
-    if (auto status = lookupLedger (ledger, context, result))
-        status.inject (result);
-
-    return result;
-}
-
-} // RPC
-} // ripple
