--- src/ripple/rpc/impl/ParseAccountIds.cpp
@@ -1,44 +1,0 @@
-//------------------------------------------------------------------------------
-/*
-    This file is part of rippled: https://github.com/ripple/rippled
-    Copyright (c) 2012-2014 Ripple Labs Inc.
-
-    Permission to use, copy, modify, and/or distribute this software for any
-    purpose  with  or without fee is hereby granted, provided that the above
-    copyright notice and this permission notice appear in all copies.
-
-    THE  SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES
-    WITH  REGARD  TO  THIS  SOFTWARE  INCLUDING  ALL  IMPLIED  WARRANTIES  OF
-    MERCHANTABILITY  AND  FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR
-    ANY  SPECIAL ,  DIRECT, INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES
-    WHATSOEVER  RESULTING  FROM  LOSS  OF USE, DATA OR PROFITS, WHETHER IN AN
-    ACTION  OF  CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF
-    OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
-*/
-//==============================================================================
-
-#include <BeastConfig.h>
-#include <ripple/rpc/impl/ParseAccountIds.h>
-
-namespace ripple {
-namespace RPC {
-
-hash_set<AccountID>
-parseAccountIds (Json::Value const& jvArray)
-{
-    hash_set<AccountID> result;
-    for (auto const& jv: jvArray)
-    {
-        if (! jv.isString())
-            return hash_set<AccountID>();
-        auto const id =
-            parseBase58<AccountID>(jv.asString());
-        if (! id)
-            return hash_set<AccountID>();
-        result.insert(*id);
-    }
-    return result;
-}
-
-} // RPC
-} // ripple
