--- src/ripple/rpc/impl/Utilities.h
@@ -1,69 +1,0 @@
-//------------------------------------------------------------------------------
-/*
-    This file is part of rippled: https://github.com/ripple/rippled
-    Copyright (c) 2012, 2013 Ripple Labs Inc.
-
-    Permission to use, copy, modify, and/or distribute this software for any
-    purpose  with  or without fee is hereby granted, provided that the above
-    copyright notice and this permission notice appear in all copies.
-
-    THE  SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES
-    WITH  REGARD  TO  THIS  SOFTWARE  INCLUDING  ALL  IMPLIED  WARRANTIES  OF
-    MERCHANTABILITY  AND  FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR
-    ANY  SPECIAL ,  DIRECT, INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES
-    WHATSOEVER  RESULTING  FROM  LOSS  OF USE, DATA OR PROFITS, WHETHER IN AN
-    ACTION  OF  CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF
-    OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
-*/
-//==============================================================================
-
-#ifndef RIPPLE_RPC_IMPL_UTILITIES_H_INCLUDED
-#define RIPPLE_RPC_IMPL_UTILITIES_H_INCLUDED
-
-#include <ripple/ledger/TxMeta.h>
-#include <ripple/rpc/impl/Tuning.h>
-#include <ripple/protocol/Seed.h>
-#include <boost/optional.hpp>
-
-namespace Json {
-class Value;
-}
-
-namespace ripple {
-
-class Transaction;
-
-namespace RPC {
-
-struct Context;
-
-void
-addPaymentDeliveredAmount (
-    Json::Value&,
-    Context&,
-    std::shared_ptr<Transaction>,
-    TxMeta::pointer);
-
-/** Inject JSON describing ledger entry
-
-    Effects:
-        Adds the JSON description of `sle` to `jv`.
-
-        If `sle` holds an account root, also adds the
-        urlgravatar field JSON if sfEmailHash is present.
-*/
-void
-injectSLE (Json::Value& jv,
-    SLE const& sle);
-
-/** Retrieve the limit value from a Context, or set a default -
-    then restrict the limit by max and min if not an ADMIN request.
-
-    If there is an error, return it as JSON. */
-boost::optional<Json::Value> readLimitField(
-    unsigned int& limit, Tuning::LimitRange const&, Context const&);
-
-} // RPC
-} // ripple
-
-#endif
